ECA-8039 Fixed integer overflow during upgrade.
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@33042 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ modules/ejbca-ejb/src/org/ejbca/core/ejb/upgrade/UpgradeSessionBean.java
@@ -1727,8 +1727,8 @@ public class UpgradeSessionBean implements UpgradeSessionLocal, UpgradeSessionRe
                 log.debug("The property ocsp.expiredcert.retentionperiod has the value '" + ConfigurationHolder.getString("ocsp.expiredcert.retentionperiod") + "'.");
             }
             final String configuredValue = ConfigurationHolder.getString("ocsp.expiredcert.retentionperiod");
-            final int retentionPeriodInSeconds = Integer.parseInt(StringUtils.trim(configuredValue));
-            if (retentionPeriodInSeconds == -1) {
+            final long retentionPeriodInSeconds = Long.parseLong(StringUtils.trim(configuredValue));
+            if (retentionPeriodInSeconds == -1L) {
                 // The archive cutoff extension is disabled
                 log.info("The archive cutoff extension is disabled on this EJBCA instance. Nothing to do.");
                 return;
@@ -1739,7 +1739,7 @@ public class UpgradeSessionBean implements UpgradeSessionLocal, UpgradeSessionRe
                 final OcspKeyBinding ocspKeyBinding = (OcspKeyBinding) internalKeyBindingDataSession.getInternalKeyBinding(ocspKeyBindingId);
                 final List<String> ocspExtensions = ocspKeyBinding.getOcspExtensions();
                 ocspExtensions.add(OcspArchiveCutoffExtension.EXTENSION_OID);
-                ocspKeyBinding.setRetentionPeriod(SimpleTime.getInstance(retentionPeriodInSeconds * 1000));
+                ocspKeyBinding.setRetentionPeriod(SimpleTime.getInstance(retentionPeriodInSeconds * 1000L));
                 ocspKeyBinding.setOcspExtensions(ocspExtensions);
                 internalKeyBindingDataSession.mergeInternalKeyBinding(ocspKeyBinding);
                 log.info("Added id-pkix-ocsp-archive-cutoff with a retention period of " + retentionPeriodInSeconds + " seconds to OCSP key binding "
